+++ CryptoPortfolio.Business.Contracts/CryptoBits/Ticker.cs
@@ -10,16 +10,16 @@ namespace CryptoPortfolio.Business.Contracts.CryptoBits
     public class Ticker
     {
         public int rank { get; set; }
-        public decimal priceUSD { get; set; }
-        public decimal priceBTC { get; set; }
+        public double priceUSD { get; set; }
+        public double priceBTC { get; set; }
         public int volume24hUSD { get; set; }
-        public decimal marketCapUSD { get; set; }
+        public double marketCapUSD { get; set; }
         public long availableSupply { get; set; }
         public long totalSupply { get; set; }
         public long maxSupply { get; set; }
-        public decimal percentChange1h { get; set; }
-        public decimal percentChange24h { get; set; }
-        public decimal percentChange7d { get; set; }
+        public double percentChange1h { get; set; }
+        public double percentChange24h { get; set; }
+        public double percentChange7d { get; set; }
         public DateTime lastUpdated { get; set; }
     }
 }
